requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
requests
beautifulsoup4
faiss-cpu
sentence-transformers
numpy
pandas
tqdm
urllib3
lxml
openai
